+++ Tools/RegisterFunctions.h
@@ -157,6 +157,6 @@ public:
     void registerFunctions(NTupleReader& tr);
 };
 
-void drawSBregionDefCopy(const double ymin_Yields = 0.05, const double ymax_Yields = 500., const bool logscale = true);
+void drawSBregionDefCopy(const float ymin_Yields = 0.05, const float ymax_Yields = 500., const bool logscale = true);
 
 #endif
