+++ Tools/systematics.C
@@ -21,7 +21,7 @@ int main(int argc, char* argv[])
     };
 
     string dataSets = "ZJetsToNuNu";
-    double lumi = AnaSamples::luminosity;
+    float lumi = AnaSamples::luminosity;
 
     while((opt = getopt_long(argc, argv, "pD:L:", long_options, &option_index)) != -1)
     {
